+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 - Fix bug in parser with floats
 - Make `date` filter work with string input in `YYYY-MM-DD` format
 - Big parsing improvement (~20-40%) for projects with macros and inheritance
+- Add `Tera::extend` to extend another instance of Tera
+- Add `Tera::full_reload` that will re-run the glob and parse all templates found.
+- Make `Tera::add_raw_template{s}` and `Tera::add_template_file{s}` part of the public API
+
 
 ## 0.8.0 (2017-03-03)